backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
google-auth
requests
python-multipart
passlib[bcrypt]
